DEV-2420: rename files.
